+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - new: add `inHoveredRangeClassName` prop to `DaySlots` component to be able to style the hovered range.
 - new: add `inHoveredRangeParentStyles` prop to `DaySlots` component to be able to style the hovered range.
 - new: add `inHoveredRangeParentClassName` prop to `DaySlots` component to be able to style the hovered range.
+- new: add `prevButtonClassName` prop to `Header` component.
+- new: add `nextButtonClassName` prop to `Header` component.
+- new: add `prevButtonStyles` prop to `Header` component.
+- new: add `nextButtonStyles` prop to `Header` component.
+- new: add `rootStyles` prop to `Header` component.
 - chore: remove warning for `unique "key" prop` and `defaultValue for select component`.
 
 ### version 1.1.7